Another SN refactor

+++ SN/TicketDecryptor.cpp
@@ -6,15 +6,15 @@
 #include "TicketDecryptor.h"
 
 TicketDecryptor::TicketDecryptor() {
-    memset(key, 0x00, CryptoPP::AES::DEFAULT_KEYLENGTH);
-    memset(iv, 0x00, CryptoPP::AES::BLOCKSIZE);
+    memset(key, 0x00, (size_t) CryptoPP::AES::DEFAULT_KEYLENGTH);
+    memset(iv, 0x00, (size_t) CryptoPP::AES::BLOCKSIZE);
 }
 
 std::string TicketDecryptor::decryptTicket(std::string ticket) {
     std::string ciphertext = ticket;
     std::string decryptedtext;
 
-    CryptoPP::AES::Decryption aesDecryption(key, CryptoPP::AES::DEFAULT_KEYLENGTH);
+    CryptoPP::AES::Decryption aesDecryption(key, (size_t) CryptoPP::AES::DEFAULT_KEYLENGTH);
     CryptoPP::CBC_Mode_ExternalCipher::Decryption cbcDecryption(aesDecryption, iv);
 
     CryptoPP::StreamTransformationFilter stfDecryptor(cbcDecryption, new CryptoPP::StringSink(decryptedtext));
